moved UserDashboard into components and removed pagination logic to fix build errors
